pyproject.toml
[tool.poetry]
name = "automation-demo"
version = "0.1.0"
description = ""
authors = ["Jonathan Aires <airesjonathan123@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
selenium = "4.13.0"
undetected-chromedriver = "^3.5.5"
python-dotenv = "^1.0.1"
nodriver = "^0.29rc2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
